added JPA repositories and Custom methods for repository's independent queries

+++ src/main/java/cr/dbo/ops/repository/COMMITRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import cr.dbo.ops.COMMIT;
 
-public interface COMMITRepository extends CrudRepository<COMMIT, Long>{
+public interface COMMITRepository extends CrudRepository<COMMIT, Integer>{
 
 }
